refactor: ナレッジベース取り込みを knowledge_materials に一本化
case search のデータ基盤（documents/chunks）を udai.y 作の knowledge_materials
モジュールに集約する。#17 で agent-first-meeting 側に重複追加した取り込みを削除し、
knowledge_materials の ingest を実インフラで動くよう修正して「正の取り込み口」にする。

- knowledge_materials/src/ingest_documents.py を実インフラ対応に修正
  - 接続を agent_first_meeting.config.settings に統一（DB=sales-agent）
  - 埋め込みを 1536 次元に（vector policy と一致）
  - chunks をベクトルポリシー付きで自動作成
  - langchain 依存を除き、依存なしの簡易チャンク分割に
- check_count.py / search_knowledge.py も settings / sales-agent / 1536 に統一
- 重複していた agent-first-meeting/scripts/ingest_documents.py と
  agent-first-meeting/data/*.pptx を削除（knowledge_materials/data を正とする）
- README / azure_setup の取り込み手順を knowledge_materials の ingest に更新

アプリの検索（agent-first-meeting/case_search.py）は従来どおり documents/chunks を
読むため、データ基盤 = knowledge_materials になる。ライブ検証で ingest→chunks 6件→
case_search が hits 返却、check_count 動作を確認済み。
